refactor : 랜덤값 생성 위치 변경 (Cars -> Car)
